Move hashPassword to a separate file because Vercel Edge runtime doesn't like hash-wasm
