+++ src/defaults.ts
@@ -1,0 +1,10 @@
+export enum Defaults {
+  Version = '001',
+  ArgonLength = 32,
+  ArgonSaltLength = 16,
+  ArgonIterations = 5,
+  ArgonMemLimit = 67108864,
+  ArgonOutputKeyBytes = 64,
+  EncryptionKeyLength = 32,
+  EncryptionNonceLength = 24,
+}
