+++ src/types.ts
@@ -1,0 +1,5 @@
+export type HexString = string
+export type Utf8String = string
+export type Base64String = string
+export type EncryptedMessage = string
+export type EncryptedBlobMessage = Uint8Array
